backend/requirements.txt
fastapi
uvicorn
langgraph
pydantic
python-dotenv
langchain_groq